+++ app/scripts/metamask-controller.js
@@ -64,6 +64,8 @@ module.exports = class MetamaskController {
       recoverSeed: idStore.recoverSeed.bind(idStore),
       // coinbase
       buyEth: this.buyEth.bind(this),
+      // shapeshift
+      createShapeShiftTx: this.createShapeShiftTx.bind(this),
     }
   }
 
@@ -317,6 +319,9 @@ module.exports = class MetamaskController {
     })
   }
 
+  createShapeShiftTx (depositAddress, depositType) {
+    this.configManager.createShapeShiftTx(depositAddress, depositType)
+  }
 }
 
 function noop () {}